Create 38.Copy List with Random Pointer.cpp
Given a linked list having two pointers in each node.
The first one points to the next node of the list,
however, the other pointer is random and can point
to any node of the list or null. The task is to create a
deep copy of the given linked list and return its head.
We will validate whether the linked list is a copy of
the original linked list or not.

A deep copy of a Linked List means we do not copy
the references of the nodes of the original Linked List
rather for each node in the original Linked List, a new
node is created.