Create 56.N Queens.cpp
You are given an integer 'N'. For a given 'N' x 'N'
chessboard, find a way to place 'N' queens such that
no queen can attack any other queen on the
chessboard.

A queen can be killed when it lies in the same row, or
same column, or the same diagonal of any of the
other queens. You have to print all such
configurations.